Code/testing_tools/evaluation_framework/requirements.txt
openai==2.17.0
tiktoken==0.11.0
python-dotenv==1.2.1
requests==2.32.5
pymupdf==1.26.7
python-docx==1.2.0
comet-ml==3.55.0
groq==0.30.0
